Adds changes to create constants file and move constants into that file.
Rename page title
