+++ src/constants.ts
@@ -1,0 +1,9 @@
+const NO_RESULTS_FOUND: string = 'No results found';
+const LOADING_MESSAGE: string = 'Fetching results...';
+const DEBOUNCE_DELAY: number = 300;
+
+export {
+    NO_RESULTS_FOUND,
+    LOADING_MESSAGE,
+    DEBOUNCE_DELAY
+}
